Added gui.py, updated main.py and brain.py
Almost an entirely new version of the Wordle Solver, I created a UI that allows the user to interact with the rest of the function by clicking on each individual letter to change it between incorrect letter, yellow, and green. Completely removes the need to use the command line. Created new function in brain.py that is used in conjunction with the new ui that allows the user to specifically chose which spots in the word a green/yellow letter occurs instead of just providing the letters and having to find the index in code (which cause some issues when double letters occurred)